FastAPI_with_Artyom_Shumeyko/requirements.txt
aioredis==2.0.1
alembic==1.10.2
amqp==5.1.1
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
Automat==22.10.0
bcrypt==4.0.1
billiard==3.6.4.0
build==0.10.0
CacheControl==0.12.11
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==2.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
crashtest==0.4.1
cryptography==37.0.4
decorator==5.1.1
distlib==0.3.6
dnspython==2.3.0
dulwich==0.21.3
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.95.0
fastapi-cache2==0.2.1
fastapi-users==10.4.1
fastapi-users-db-sqlalchemy==5.0.0
filelock==3.8.0
flower==1.2.0
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
humanize==4.6.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.16.0
keyring==23.9.3
kombu==5.2.4
lockfile==0.12.2
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.2
more-itertools==8.14.0
msgpack==1.0.4
orjson==3.8.9
outcome==1.2.0
packaging==22.0
passlib==1.7.4
pendulum==2.1.2
pexpect==4.8.0
pipenv==2023.3.20
pkginfo==1.9.6
platformdirs==2.5.2
pluggy==1.0.0
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pyproject_hooks==1.0.0
pyrsistent==0.18.1
pytest==7.3.0
pytest-asyncio==0.21.0
pytest-tornasync==0.6.0.post2
pytest-trio==0.8.0
pytest-twisted==1.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
rapidfuzz==2.13.7
redis==4.5.4
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
SecretStorage==3.3.3
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.8
starlette==0.26.1
tomli==2.0.1
tomlkit==0.11.5
tornado==6.2
trio==0.22.0
trove-classifiers==2022.12.22
Twisted==22.10.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.12
uvicorn==0.21.1
uvloop==0.17.0
vine==5.0.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
watchfiles==0.19.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0
zope.interface==6.0
gunicorn
flower
